LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cross-origin-isolated-permission.https-expected.txt


FAIL frame: origin = https://localhost:9443, value = undefined assert_equals: expected (boolean) true but got (undefined) undefined
FAIL frame: origin = https://localhost:9443, value = * assert_equals: expected (boolean) true but got (undefined) undefined
FAIL frame: origin = https://localhost:9443, value = self assert_equals: expected (boolean) true but got (undefined) undefined
FAIL frame: origin = https://localhost:9443, value = (\) assert_equals: expected (boolean) false but got (undefined) undefined
FAIL frame: origin = https://127.0.0.1:9443, value = undefined assert_equals: expected (boolean) false but got (undefined) undefined
FAIL frame: origin = https://127.0.0.1:9443, value = * assert_equals: expected (boolean) false but got (undefined) undefined
FAIL frame: origin = https://127.0.0.1:9443, value = self assert_equals: expected (boolean) false but got (undefined) undefined
FAIL frame: origin = https://127.0.0.1:9443, value = (\) assert_equals: expected (boolean) false but got (undefined) undefined
FAIL dedicated worker: scheme = https, value = undefined assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = https, value = * assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = https, value = self assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = https, value = (\) assert_equals: expected (boolean) false but got (undefined) undefined
FAIL dedicated worker: scheme = data, value = undefined The operation is insecure.
FAIL dedicated worker: scheme = data, value = * The operation is insecure.
FAIL dedicated worker: scheme = data, value = self The operation is insecure.
FAIL dedicated worker: scheme = data, value = (\) The operation is insecure.
FAIL dedicated worker: scheme = blob, value = undefined assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = blob, value = * assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = blob, value = self assert_equals: expected (boolean) true but got (undefined) undefined
FAIL dedicated worker: scheme = blob, value = (\) assert_equals: expected (boolean) false but got (undefined) undefined
FAIL shared worker: withCoopCoep = false Can't find variable: SharedWorker
FAIL shared worker: withCoopCoep = true Can't find variable: SharedWorker
FAIL service worker: withCoopCoep = false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL service worker: withCoopCoep = true assert_equals: expected (boolean) true but got (undefined) undefined
